requirements.txt
streamlit
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
pillow
numpy==1.24.4
opencv-python-headless==4.8.1.78
matplotlib
